Added separate min_api_level
minSdkVersion is one of variables that are nearly always set to value lower
than targetSdkVersion, but present because otherwise it defaults to 1. It also
usually does not change during project, so it makes sense to set it up during
project creation. This variable is also included by the new Android build
scripts:

http://tools.android.com/tech-docs/new-build-system/user-guide#TOC-Basic-Build-Customization