Do not prefer Idea editor
I've seen forks of this project that only change editor from Idea to
other editor in the plugins.sbt file. I believe it's best to not include
default IDE, but to include in README file how to work with each popular
IDE that work with Scala. It will help centralize the effort to create one,
single standard project template, and thus better one. The list here is taken
from description of Scala on Typesafe homepage.