Disable forced quit
Forced quit is workaround for apps running on old versions of MacOS X inside
Java Webstart. It quits the application by harsh System.exit(-1), that is
caught by SBT as exception.

I believe it is better to disable it during development (development even if
done on Mac, isn't done inside Webstart), making the application close nicely.
It means, that if you see exception on console, something is really wrong
with your code, not with Webstart on MacOS X.